gateway/requirements.txt
httpx==0.25.2
pydantic==2.5.0
python-multipart==0.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
aiofiles==23.2.1
python-dotenv==1.0.0
sqlalchemy==2.0.27
supabase==2.3.4
psycopg2-binary==2.9.10
fastapi==0.111.0
uvicorn[standard]==0.30.0
